Add files via upload
## Adidas-USA-Sales-A-Statistical-and-Machine-Learning-Perspective
Data analysis of Adidas USA sales, incl. stat analysis, visualization, and a predictive model. Dataset, code, visualizations, and reports included. Code and trained model files for predictive model also included. Materials provided to reproduce and understand the analysis.
This repository contains a data analysis of Adidas USA sales. 
The project includes statistical analysis, visualization, and the development of a predictive model. 
The dataset used for the analysis is included, as well as Jupyter notebooks with the code for the analysis and model building. 
The repository also includes visualizations and reports summarizing the findings of the analysis, as well as the code and trained model files for the predictive model. 
All materials in the repository are provided to reproduce and understand the analysis.